Aggiunto la funzione makewall che crea dei muri in giro per la mappa sapendo le coordinate degli estremi(si dà per scontato che gli estremi inseriti siano all'interno della mappa) e commentate ci sono delle righe che consistono in una prova del funzionamento
